+++ src/clientapp/src/app/manage-targets-editor/manage-targets-editor.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { PeerGroup, ServerValidationError, Target, WebapiService } from '../webapi.service';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { AppSharedModule } from '../app-shared.module';
@@ -11,6 +11,7 @@ import { ValidationErrorsDisplayComponent } from '../validation-errors-display/v
   selector: 'app-manage-targets-editor',
   standalone: true,
   imports: [CommonModule, FormsModule, AppSharedModule, ValidationErrorsDisplayComponent],
+  providers: [ConfirmationService, MessageService],
   templateUrl: './manage-targets-editor.component.html',
   styleUrl: './manage-targets-editor.component.scss'
 })
@@ -36,7 +37,9 @@ export class ManageTargetsEditorComponent {
 
   @Output() onFinish = new EventEmitter<boolean>();
 
-  constructor(private messageService: MessageService, private webapiService: WebapiService) { }
+  constructor(private messageService: MessageService, 
+    private confirmationService: ConfirmationService,
+    private webapiService: WebapiService) { }
 
   getLookupData() {
     this.webapiService.getPeerGroupList().subscribe(lookup => {
@@ -64,24 +67,38 @@ export class ManageTargetsEditorComponent {
       });
   }
 
-  delete(){
+  delete(event: Event) {
     if (!this.target) {
       return;
     }
-    this.webapiService.deleteTarget(this.target)
-      .subscribe({
-        next: data => {
-        },
-        error: error => {
-          let response = error as HttpErrorResponse;
-          if (response) {
-            this.validationResult = response.error as ServerValidationError;
-          }
-        },
-        complete: () => {
-          this.onFinish.emit(true);
-        },
-      });
+
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      acceptIcon: "none",
+      rejectIcon: "none",
+      rejectButtonStyleClass: "p-button-text",
+      accept: () => {
+        this.webapiService.deleteTarget(this.target)
+          .subscribe({
+            next: data => {
+            },
+            error: error => {
+              let response = error as HttpErrorResponse;
+              if (response) {
+                this.validationResult = response.error as ServerValidationError;
+              }
+            },
+            complete: () => {
+              this.onFinish.emit(true);
+            },
+          });
+      },
+      reject: () => {
+      }
+    });
   }
 
   cancel() {
